user-service/requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
django-cors-headers==4.3.1
cryptography==41.0.7
requests==2.31.0 #lets user service call auth service to verify tokens!